@@ -36,62 +36,63 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/forms": "7.2.15",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "@ngrx/effects": "7.4.0",
-    "@ngrx/router-store": "7.4.0",
-    "@ngrx/store": "7.4.0",
-    "@nrwl/nx": "7.8.6",
+    "@angular/animations": "8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "8.2.14",
+    "@angular/compiler": "8.2.14",
+    "@angular/core": "8.2.14",
+    "@angular/forms": "8.2.14",
+    "@angular/platform-browser": "8.2.14",
+    "@angular/platform-browser-dynamic": "8.2.14",
+    "@angular/router": "8.2.14",
+    "@ngrx/effects": "8.6.0",
+    "@ngrx/router-store": "8.6.0",
+    "@ngrx/store": "8.6.0",
+    "@nrwl/nx": "7.8.7",
     "core-js": "^2.5.4",
-    "rxjs": "^6.5.2",
-    "zone.js": "^0.8.26"
+    "rxjs": "^6.5.4",
+    "zone.js": "^0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular-devkit/build-ng-packagr": "~0.13.1",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "7.2.15",
-    "@angular/language-service": "7.2.15",
-    "@ngrx/schematics": "7.4.0",
-    "@ngrx/store-devtools": "7.4.0",
-    "@nrwl/builders": "7.5.2",
-    "@nrwl/schematics": "7.5.2",
-    "@types/jasmine": "~3.3.9",
-    "@types/jasminewd2": "~2.0.5",
-    "@types/jest": "^23.0.0",
-    "@types/node": "~11.9.4",
+    "@angular-devkit/build-angular": "^0.803.24",
+    "@angular-devkit/build-ng-packagr": "^0.803.24",
+    "@angular/cli": "8.3.24",
+    "@angular/compiler-cli": "8.2.14",
+    "@angular/language-service": "8.2.14",
+    "@ngrx/schematics": "8.6.0",
+    "@ngrx/store-devtools": "8.6.0",
+    "@nrwl/builders": "7.8.7",
+    "@nrwl/schematics": "8.12.1",
+    "@types/jasmine": "^3.5.2",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/jest": "^25.1.1",
+    "@types/node": "^13.7.0",
     "codecov.io": "^0.1.6",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~3.3.0",
-    "jasmine-marbles": "0.4.1",
+    "codelyzer": "~5.2.1",
+    "jasmine-core": "~3.5.0",
+    "jasmine-marbles": "0.6.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "jest": "^23.0.0",
-    "jest-preset-angular": "6.0.1",
-    "karma": "~4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage": "^1.1.2",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^4.4.0",
+    "ts-jest": "^24.0.0",
+    "jest": "^24.1.0",
+    "jest-preset-angular": "7.0.0",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "^2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.1.1",
+    "karma-jasmine": "~3.1.0",
+    "karma-jasmine-html-reporter": "^1.5.1",
+    "ng-packagr": "^5.7.1",
     "ngrx-store-freeze": "0.2.4",
-    "prettier": "1.16.4",
-    "pretty-quick": "^1.10.0",
-    "protractor": "~5.4.0",
-    "remark-cli": "^6.0.1",
-    "remark-lint": "^6.0.4",
-    "remark-preset-lint-recommended": "^3.0.2",
-    "ts-node": "~8.0.2",
-    "tsickle": "0.34.3",
-    "tslib": "^1.9.0",
-    "tslint": "~5.12.1",
-    "typescript": "~3.2.4"
+    "prettier": "1.19.1",
+    "pretty-quick": "^2.0.1",
+    "protractor": "^5.4.3",
+    "remark-cli": "^7.0.1",
+    "remark-lint": "^6.0.5",
+    "remark-preset-lint-recommended": "^3.0.3",
+    "ts-node": "^8.6.2",
+    "tsickle": "0.38.0",
+    "tslib": "^1.10.0",
+    "tslint": "~6.0.0",
+    "typescript": "~3.4.0"
   }
 }
