@@ -1,7 +1,7 @@
 module.exports = {
   testMatch: ['**/+(*.)+(spec|test).+(ts|js)?(x)'],
   transform: {
-    '^.+\\.(ts|js|html)$': 'jest-preset-angular/preprocessor.js'
+    '^.+\\.(ts|js|html)$': 'ts-jest'
   },
   resolver: '@nrwl/builders/plugins/jest/resolver',
   moduleFileExtensions: ['ts', 'js', 'html'],
